convert setLogEventNum to updateLogEventNum
