+++ src/components/AppController.tsx
@@ -3,24 +3,15 @@ import React, {
     useRef,
 } from "react";
 
-import useLogFileManagerStore from "../stores/logFileManagerProxyStore";
 import useLogFileStore from "../stores/logFileStore";
-import {handleErrorWithNotification} from "../stores/notificationStore";
 import useQueryStore from "../stores/queryStore";
-import useUiStore from "../stores/uiStore";
 import useViewStore from "../stores/viewStore";
 import {Nullable} from "../typings/common";
-import {UI_STATE} from "../typings/states";
 import {UrlHashParams} from "../typings/url";
 import {
     CURSOR_CODE,
     CursorType,
 } from "../typings/worker";
-import {
-    findNearestLessThanOrEqualElement,
-    isWithinBounds,
-} from "../utils/data";
-import {clamp} from "../utils/math";
 import {
     getWindowUrlHashParams,
     getWindowUrlSearchParams,
@@ -31,64 +22,23 @@ import {
 
 
 /**
- * Updates the log event number in the URL to `logEventNum` if it's within the bounds of
- * `logEventNumsOnPage`.
- *
- * @param logEventNum
- * @param logEventNumsOnPage
- * @return Whether `logEventNum` is within the bounds of `logEventNumsOnPage`.
- */
-const updateUrlIfEventOnPage = (
-    logEventNum: number,
-    logEventNumsOnPage: number[]
-): boolean => {
-    if (false === isWithinBounds(logEventNumsOnPage, logEventNum)) {
-        return false;
-    }
-
-    const nearestIdx = findNearestLessThanOrEqualElement(
-        logEventNumsOnPage,
-        logEventNum
-    );
-
-    // Since `isWithinBounds` returned `true`, then:
-    // - `logEventNumsOnPage` must bound `logEventNum`.
-    // - `logEventNumsOnPage` cannot be empty.
-    // - `nearestIdx` cannot be `null`.
-    //
-    // Therefore, we can safely cast:
-    // - `nearestIdx` from `Nullable<number>` to `number`.
-    // - `logEventNumsOnPage[nearestIdx]` from `number | undefined` to `number`.
-    const nearestLogEventNum = logEventNumsOnPage[nearestIdx as number] as number;
-
-    updateWindowUrlHashParams({
-        logEventNum: nearestLogEventNum,
-    });
-
-    return true;
-};
-
-/**
  * Updates view-related parameters from URL hash.
- *
- * @param hashParams
  */
-const updateViewHashParams = (hashParams: UrlHashParams): void => {
-    const {isPrettified, logEventNum} = hashParams;
-    const {updateIsPrettified, setLogEventNum} = useViewStore.getState();
+const updateViewHashParams = (): void => {
+    const {isPrettified, logEventNum} = getWindowUrlHashParams();
+    const {updateIsPrettified, updateLogEventNum} = useViewStore.getState();
 
     updateIsPrettified(isPrettified);
-    setLogEventNum(logEventNum);
+    updateLogEventNum(logEventNum);
 };
 
 /**
  * Updates query-related parameters from URL hash.
  *
- * @param hashParams
  * @return Whether any query parameters were modified.
  */
-const updateQueryHashParams = (hashParams: UrlHashParams): boolean => {
-    const {queryIsCaseSensitive, queryIsRegex, queryString} = hashParams;
+const updateQueryHashParams = (): boolean => {
+    const {queryIsCaseSensitive, queryIsRegex, queryString} = getWindowUrlHashParams();
     const {
         queryIsCaseSensitive: currentQueryIsCaseSensitive,
         queryIsRegex: currentQueryIsRegex,
@@ -118,9 +68,8 @@ const updateQueryHashParams = (hashParams: UrlHashParams): boolean => {
  * @return The parsed URL hash parameters.
  */
 const handleHashChange = (ev: Nullable<HashChangeEvent>): UrlHashParams => {
-    const hashParams = getWindowUrlHashParams();
-    updateViewHashParams(hashParams);
-    const isQueryModified = updateQueryHashParams(hashParams);
+    updateViewHashParams();
+    const isQueryModified = updateQueryHashParams();
     const isTriggeredByHashChange = null !== ev;
     if (isTriggeredByHashChange && isQueryModified) {
         const {startQuery} = useQueryStore.getState();
@@ -128,6 +77,7 @@ const handleHashChange = (ev: Nullable<HashChangeEvent>): UrlHashParams => {
     }
 
     // Remove empty or falsy parameters.
+    const hashParams = getWindowUrlHashParams();
     updateWindowUrlHashParams({
         ...hashParams,
     });
@@ -147,9 +97,6 @@ interface AppControllerProps {
  * @return
  */
 const AppController = ({children}: AppControllerProps) => {
-    // States
-    const logEventNum = useViewStore((state) => state.logEventNum);
-
     // Refs
     const isInitialized = useRef<boolean>(false);
 
@@ -183,38 +130,6 @@ const AppController = ({children}: AppControllerProps) => {
             window.removeEventListener("hashchange", handleHashChange);
         };
     }, []);
-
-    // On `logEventNum` update, clamp it then switch page if necessary or simply update the URL.
-    useEffect(() => {
-        const {numEvents} = useLogFileStore.getState();
-        if (0 === numEvents || URL_HASH_PARAMS_DEFAULT.logEventNum === logEventNum) {
-            return;
-        }
-
-        const clampedLogEventNum = clamp(logEventNum, 1, numEvents);
-        const {beginLineNumToLogEventNum} = useViewStore.getState();
-        const logEventNumsOnPage: number [] = Array.from(beginLineNumToLogEventNum.values());
-        if (updateUrlIfEventOnPage(clampedLogEventNum, logEventNumsOnPage)) {
-            // No need to request a new page since the log event is on the current page.
-            return;
-        }
-
-        // If the log event is not on the current page, request a new page.
-        const {setUiState} = useUiStore.getState();
-        setUiState(UI_STATE.FAST_LOADING);
-        (async () => {
-            const {logFileManagerProxy} = useLogFileManagerStore.getState();
-            const cursor: CursorType = {
-                code: CURSOR_CODE.EVENT_NUM,
-                args: {eventNum: clampedLogEventNum},
-            };
-            const {isPrettified} = useViewStore.getState();
-
-            const pageData = await logFileManagerProxy.loadPage(cursor, isPrettified);
-            const {updatePageData} = useViewStore.getState();
-            updatePageData(pageData);
-        })().catch(handleErrorWithNotification);
-    }, [logEventNum]);
 
     return children;
 };
